model/requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
openai>=1.0.0  # For chat completions
numpy>=1.21.0
pandas>=1.3.0  # For data handling
scikit-learn>=0.24.0  # For text processing
python-Levenshtein>=0.12.0  # For string matching
# Hugging Face dependencies
transformers>=4.30.0
torch>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
accelerate>=0.20.0
sentencepiece>=0.1.99
sentence-transformers>=2.2.0  # For embeddings